jmasale/requirements.txt
asgiref==3.11.1
Django==6.0.3
celery==5.5.3
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
